backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["VishalPokharel <077bct094.vishal@pcampus.edu.np>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
langchain = "^0.3.10"
qdrant-client = "^1.12.1"
python-decouple = "^3.8"
langchain-huggingface = "^0.1.2"
langchain-community = "^0.3.10"
langchain-qdrant = "^0.2.0"
beautifulsoup4 = "^4.12.3"
langchain-groq = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
